Add files via upload
Hello, I present the solution to the OM1 challenge.  The obstacle avoidance code is very similar to the one used in solving the challenge in webinar # 3.  This time I have added a GS sensor reading module (ReadGroudSensors).  I normalized the values measured by each sensor between 0 and 1000 and reversed the logic.  I estimated the average value (Position) and the position error (ErrorPosition).  I also added memory to know from which side the line was detected in case of losing it and thus stop the robot and start the corresponding recovery maneuver.

 The part that was most difficult for me to control was the turning of the fork.  I managed to control the turn by adding a routine that fires upon detection of the crossover, in this case the left side.