Line Follower Robot V4 - On Webots
Line Follower Robot V4 - On Webots
This is the virtual version of the Line Follower Robot V4 that I made physically controlled by Arduino (https://youtu.be/n3PUQIn_-08), now made in the free Webots application (https://cyberbotics.com).

The robot structure was created in TinkeCAD (https://www.tinkercad.com), they were imported in OBJ format, to be converted into WBO format and to be used as part of the design. With these files and the different objects I made the structure of the robot to simulate the mechanical operation.

The controller was made in C language, adapting the original Arduino code to make the robot work in a similar way to the physical version.

In Webots I made a surface in which I have loaded different tracks in which the design of the robot was put to the test.

During the design process I ran into a series of problems that I managed to solve mostly by accessing the user manual of the Webots platform (https://cyberbotics.com/doc/reference/index) and some videos on YouTube. I understand that it is a bit complex because in addition to the logical part, it is necessary to understand physical concepts related to mechanics, collisions, force, weight, gravity, lighting, etc.

I will make additional videos where I will explain how to build and design this robot in Webots, in case anyone is interested in knowing what little I have learned and that personally I find it very interesting since it gives the possibility of testing a prototype before proceeding to build physically.