Speed Line Follower Robot V4 on Webots
This is a latest version (V2) with better performance. I set the ERP = 1 (Error Reduction Parameter) to avoid mechanical errors in the joints of the motors and the rolling wheel. So far I have not used the supervisor API yet. I also managed to get the LEDs to work acceptably, using SpotLights directed towards the ground. In general this version works much better than the previous one, but I would like to make it progress faster (> 3500)