Universal Robot Sample V4 in Webots
This is version 4 of a sample using Universal Robot UR5e on Webots with Python and OpenCV.

This time I have added to the previous version the implementation of an HMI and a control gate. I have applied additional concepts such as the use of displays, mouse events, handling of images to change textures. As a reference I have used the official Webots manual and several of the examples included in the application.