MecanumR
This is an example of using Mecanum wheels. I use 3D obj models. The controller is in Python, and I use an Xbox joystick to control the robot. The controller uses inverse kinematics to determine the angular velocities of each wheel, using the linear and angular velocities determined by the joystick